Update README.md
Great plugin, exactly what I looked for! The import was missing, especially the Dio import. It costed me 30 mins to figure this out and a little bit frust, because Android Studio is not automatically hinting to import this lib. I think that would make the plugin more popular since then it is a working out of the box sample!